NYC Bike Usage Tableau work uploaded
NYC Bike Trip Analysis

The data provided included the whole year 2017 through the months from January to December. This analysis is based in the data behavior of the 2017 year only. 

Considerations:
The field Trip Duration in Hours was created based in Trip Duration in seconds converted to hours to make calculation in hours. 
The field Month was created based on the Start Time -datetime- field to make calculations in months.
Age was calculated based on Birth year field. 

Analysis
•	Male is the gender that use bike services the most through the whole year: 322,368. Female is the second place with 138,299 users. Unknown third. See Trip Duration vs. Gender graph.
•	August is the month with more users that use this service. See Trip vs. Gender graph. 
•	The Male gender and Age 29 followed by Female gender and Age 28 are the groups of users with more trips. A curious fact is that the older age is a male 160 user. The older female is age 132. See the Age Gender vs. Number Records worksheet. Probably there are errors in the data entry Birth Year field of the data provided by the Citi program. 
•	 The Male and 31 age population make the most duration trip in the month of October. See the Age Gender vs. Duration trip Table. 
•	The Start Station name “Pershing Square South” is the most used by renters, as we can see at the Location vs. Trips map.
•	Subscriber user type and Male are the most users in this program -almost 75%-. Subscriber and Female represent almost 25% as we can see in the User Type Vs. Usage pie chart.
•	The Pershing Square North is the top of the Top 10 Stations to Start a trip and it is also the top of the Top 10 Stations to End a trip.
•	The “13 St & 34th Ave” is the top bottom station to start a trip. See the Top 10 Start Bottom bar graph.
•	The “35 Ave & 10 St” is the top bottom station to end a trip. See the Top 10 End Bottom bar graph.

Criticism
This analysis was based in the 2017 data. It would be interesting is the CitiBike program provided a file holding all of years since the beginning in 2013, but now it is too fragmented.